files/requirements.txt
aiogram==3.12.0
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-docx==1.1.2
python-dotenv==1.0.1
